[fix] ユーザー詳細画面及びユーザー編集画面のレイアウトを修正

+++ app/controllers/favorites_controller.rb
@@ -5,15 +5,15 @@ class FavoritesController < ApplicationController
       favorite = current_user.favorites.new(post_id: @post.id)
       favorite.save
       @post.create_notification_favorite!(current_user)
-    else
-      redirect_to new_user_session_path
     end
   end
 
   def destroy
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.find_by(post_id: @post.id)
-    favorite.destroy
+    if @post.user != current_user
+      favorite = current_user.favorites.find_by(post_id: @post.id)
+      favorite.destroy
+    end
   end
 
 end
